// src/common/errors/error-code.enum.ts
export enum ErrorCode {
  INTERNAL_SERVER_ERROR = 'INTERNAL_SERVER_ERROR',
  BAD_REQUEST = 'BAD_REQUEST',
  VALIDATION_FAILED = 'VALIDATION_FAILED',
  UNAUTHORIZED = 'UNAUTHORIZED',
  FORBIDDEN = 'FORBIDDEN',
  RESOURCE_NOT_FOUND = 'RESOURCE_NOT_FOUND',
  CONFLICT = 'CONFLICT',
  TENANT_ORGANIZATION_REQUIRED = 'TENANT_ORGANIZATION_REQUIRED',
  TENANT_ORGANIZATION_FORBIDDEN = 'TENANT_ORGANIZATION_FORBIDDEN',
  INVALID_SORT_FIELD = 'INVALID_SORT_FIELD',
}